+++ src/main/java/pl/maciejak/charity/repository/InstitutionRepository.java
@@ -3,5 +3,5 @@ package pl.maciejak.charity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.maciejak.charity.entity.Institution;
 
-public interface InstitutionRepository extends JpaRepository<Institution, Integer> {
+public interface InstitutionRepository extends JpaRepository<Institution, Long> {
 }
